leaf/requirements.txt
Flask 
Flask-WTF 
pickle-mixin 
Pillow 
numpy 
tensorflow
scikit-learn
scikit-image
requests
beautifulsoup4
transformers
tf-keras
cohere
matplotlib